Implemented greedy meshing
Right now the code is verbose, and a little messy. The performance of the meshing algorithm is also only around 20% better than the per block meshing, however, there is far more room for optimization. There are many inefficient loops/memory allocations. It currently takes ~2.4 ms to generate the greedy quads for a 32x32x32 chunk. With optimizations, sub 1 ms times should be possible.